RobustNeuralNetworks.jl Documentation

Bringing robust machine learning tools to Julia.

RobustNeuralNetworks.jl Documentation

Types

Functions

Examples

Index

Docstrings

TODO:

Types

RobustNeuralNetworks.DirectParams — Type

mutable struct DirectParams{T}

Direct (implicit) parameters used to construct a REN.

Functions

Missing docstring.

Missing docstring for hmatrix_to_explicit. Check Documenter's build log for details.

Examples

Here's an example. You should not be able to see the import statement of Random.

using Random
a = 1
b = 2*rand()
2a + b

3.5465585705219635

Can we continue using variables from this example?

println(a+b)

2.5465585705219635

We can even make things look like the REPL.

julia> a = 11
julia> b = 22
julia> a + b3

We can even delay execution of an example over a few different example blocks. Start a for loop here...

for i in 1:3
    j = i^2

Then write something insightful and finish it below...

    println(j)
end

1
4
9

It's worth having a look at the @setup macro as well when you can. It will make it much easier to write examples that include a number of lines of setup which should be hidden. Having said that, it might be useful to show the reader how you set up the example!

Most of your examples should be written with the @jldoctest macro. I'll give it a go below, but have a look at how ControlSystems.jl does things too.

Example:

using Random
using RobustNeuralNetworks

batches = 50
nu, nx, nv, ny = 4, 10, 20, 2

contracting_ren_ps = ContractingRENParams{Float64}(nu, nx, nv, ny)
contracting_ren = REN(contracting_ren_ps)

x0 = init_states(contracting_ren, batches)
u0 = randn(contracting_ren.nu, batches)

x1, y1 = contracting_ren(x0, u0)  # Evaluates the REN over one timestep

println(size(y1))

# output

(2, 50)

Index

RobustNeuralNetworks.AbstractLBDN

RobustNeuralNetworks.AbstractREN

RobustNeuralNetworks.AbstractREN

RobustNeuralNetworks.AbstractRENParams

RobustNeuralNetworks.ContractingRENParams

RobustNeuralNetworks.ContractingRENParams

RobustNeuralNetworks.ContractingRENParams

RobustNeuralNetworks.DiffREN

RobustNeuralNetworks.DiffREN

RobustNeuralNetworks.DiffREN

RobustNeuralNetworks.DirectParams

RobustNeuralNetworks.DirectParams

RobustNeuralNetworks.ExplicitParams

RobustNeuralNetworks.GeneralRENParams

RobustNeuralNetworks.GeneralRENParams

RobustNeuralNetworks.LBFN

RobustNeuralNetworks.LBFN

RobustNeuralNetworks.LipschitzRENParams

RobustNeuralNetworks.LipschitzRENParams

RobustNeuralNetworks.PassiveRENParams

RobustNeuralNetworks.PassiveRENParams

RobustNeuralNetworks.REN

RobustNeuralNetworks.REN

RobustNeuralNetworks.WrapREN

RobustNeuralNetworks.WrapREN

RobustNeuralNetworks.WrapREN2

RobustNeuralNetworks.WrapREN2

RobustNeuralNetworks.WrapREN2

RobustNeuralNetworks.direct_to_explicit

RobustNeuralNetworks.direct_to_explicit

RobustNeuralNetworks.direct_to_explicit

RobustNeuralNetworks.direct_to_explicit

RobustNeuralNetworks.init_states

RobustNeuralNetworks.init_states

RobustNeuralNetworks.set_output_zero!

RobustNeuralNetworks.set_output_zero!

RobustNeuralNetworks.update_explicit!

Docstrings

Work on the presentation of this a bit....

RobustNeuralNetworks.AbstractLBDN — Type

abstract type AbstractLBDN

RobustNeuralNetworks.AbstractREN — Type

abstract type AbstractREN

RobustNeuralNetworks.AbstractREN — Method

(m::AbstractREN)(xt::VecOrMat, ut::VecOrMat)

Call the REN given internal states xt and inputs ut. If function arguments are matrices, each column must be a vector of states or inputs (allows batch simulations).

RobustNeuralNetworks.AbstractRENParams — Type

abstract type AbstractRENParams{T}

RobustNeuralNetworks.ContractingRENParams — Type

mutable struct ContractingRENParams{T} <: AbstractRENParams{T}

Parameter struct to build a contracting acyclic REN.

RobustNeuralNetworks.ContractingRENParams — Method

ContractingRENParams(nu, nx, nv, ny; ...)

Main constructor for ContractingRENParams. ᾱ ∈ (0,1] is the upper bound on contraction rate.

RobustNeuralNetworks.ContractingRENParams — Method

ContractingRENParams(nv, A, B, C, D; ...)

Alternative constructor for ContractingRENParams that initialises the REN from a stable discrete-time linear system ss(A,B,C,D).

TODO: Make compatible with αbar ≠ 1.0

RobustNeuralNetworks.DiffREN — Type

mutable struct DiffREN <: AbstractREN

Wrapper for Recurrent Equilibrium Network type which automatically re-computes explicit parameters every time the model is called.

Compatible with Flux.jl

RobustNeuralNetworks.DiffREN — Method

(m::DiffREN)(xt::VecOrMat, ut::VecOrMat)

Call the REN given internal states xt and inputs ut. If function arguments are matrices, each column must be a vector of states or inputs (allows batch simulations).

Computes explicit parameterisation each time. This may be slow if called many times!

RobustNeuralNetworks.DiffREN — Method

DiffREN(ps::AbstractRENParams)

Construct DiffREN wrapper from direct parameterisation

RobustNeuralNetworks.DirectParams — Method

DirectParams{T}(nu, nx, nv; ...)

Constructor for DirectParams struct. Allows for the following initialisation methods, specified as symbols by init argument:

:random: Random sampling for all parameters

:cholesky: Compute X with cholesky factorisation of H, sets E,F,P = I

Option D22_free specifies whether or not to train D22 as a free parameter, or constructed separately from X3, Y3, Z3. Typically use D22_free = true for a contracting REN. Default is D22_free = false.

Option D22_zero fixes D22 = 0 to remove any feedthrough. Default false.

RobustNeuralNetworks.ExplicitParams — Type

mutable struct ExplicitParams{T}

Struct containing explicit REN parameters

RobustNeuralNetworks.GeneralRENParams — Type

mutable struct GeneralRENParams{T} <: AbstractRENParams{T}

Parameter struct to build an acyclic REN with behavioural constraints encoded in Q, S, R matrices

RobustNeuralNetworks.GeneralRENParams — Method

GeneralRENParams(nu, nx, nv, ny; ...)

Main constructor for GeneralRENParams. ᾱ ∈ (0,1] is the upper bound on contraction rate.

RobustNeuralNetworks.LBFN — Type

mutable struct LBFN{T} <: AbstractLBDN

Direct parameters for an LBFN

RobustNeuralNetworks.LBFN — Method

LBFN{T}(nu, nh, ny, γ; ...)

Constructor for an LBFN with nu inputs, nv outputs, and nh = [nh1, nh2,...] specifying the size of hidden layers. User-imposed Lipschitz bound γ has a default of 1.

RobustNeuralNetworks.LipschitzRENParams — Type

mutable struct LipschitzRENParams{T} <: AbstractRENParams{T}

Parameter struct to build an acyclic REN with a guaranteed Lipschitz bound of γ ∈ ℝ

RobustNeuralNetworks.LipschitzRENParams — Method

LipschitzRENParams(nu, nx, nv, ny; ...)

Main constructor for LipschitzRENParams. ᾱ ∈ (0,1] is the upper bound on contraction rate.

RobustNeuralNetworks.PassiveRENParams — Type

mutable struct PassiveRENParams{T} <: AbstractRENParams{T}

Parameter struct to build a passive REN where Q = 0, R = 0, S = I

RobustNeuralNetworks.PassiveRENParams — Method

PassiveRENParams(nu, nx, nv, ny; ...)

Main constructor for PassiveRENParams. ᾱ ∈ (0,1] is the upper bound on contraction rate. ν>0 for incrementally strictly input passive model; v == 0 for incrementally passive model.

RobustNeuralNetworks.REN — Type

mutable struct REN <: AbstractREN

Main Recurrent Equilibrium Network type

RobustNeuralNetworks.REN — Method

REN(ps::AbstractRENParams)

Construct REN from direct parameterisation

RobustNeuralNetworks.WrapREN — Type

mutable struct WrapREN <: AbstractREN

Wrapper for Recurrent Equilibrium Network type combining direct parameters and explicit model into one type.

Requires user to manually update explicit params when direct params are changed. Not compatible with Flux.jl

RobustNeuralNetworks.WrapREN — Method

WrapREN(ps::AbstractRENParams)

Construct REN wrapper from direct parameterisation

RobustNeuralNetworks.WrapREN2 — Type

mutable struct WrapREN2 <: AbstractREN

Wrapper for Recurrent Equilibrium Network type which automatically re-computes explicit parameters whenever the direct parameters are edited.

Not compatible with Flux.jl

RobustNeuralNetworks.WrapREN2 — Method

(m::WrapREN2)(xt::VecOrMat, ut::VecOrMat)

Call the REN given internal states xt and inputs ut. If function arguments are matrices, each column must be a vector of states or inputs (allows batch simulations).

Updates the explicit parameterisation if direct parameters have been updated.

RobustNeuralNetworks.WrapREN2 — Method

WrapREN2(ps::AbstractRENParams)

Construct WrapREN2 wrapper from direct parameterisation

RobustNeuralNetworks.direct_to_explicit — Method

direct_to_explicit(ps::ContractingRENParams)

Convert direct REN parameterisation to explicit parameterisation for contracting REN

RobustNeuralNetworks.direct_to_explicit — Method

direct_to_explicit(ps::GeneralRENParams)

Convert direct REN parameterisation to explicit parameterisation using behavioural constraints encoded in Q, S, R

RobustNeuralNetworks.direct_to_explicit — Method

direct_to_explicit(ps::LipschitzRENParams)

Convert direct REN parameterisation to explicit parameterisation using Lipschitz bounded behavioural constraint

RobustNeuralNetworks.direct_to_explicit — Method

direct_to_explicit(ps::PassiveRENParams)

Convert direct REN parameterisation to explicit parameterisation using passive behavioural constraints encoded in Q, S, R

RobustNeuralNetworks.init_states — Method

init_states(m::AbstractREN, nbatches; rng=nothing)

Return matrix of (nbatches) state vectors initialised as zeros

RobustNeuralNetworks.init_states — Method

init_states(m::AbstractREN; rng=nothing)

Return state vector initialised as zeros

RobustNeuralNetworks.set_output_zero! — Method

set_output_zero!(m::AbstractRENParams)

Set output map of REN to zero

RobustNeuralNetworks.set_output_zero! — Method

set_output_zero!(m::AbstractREN)

Set output map of REN to zero

RobustNeuralNetworks.update_explicit! — Method

update_explicit!(m::WrapREN)

Update explicit model using the current direct parameters

TODO:

Add a logo

Fill out this main documentation page

See ControlSystems.jl for a good example of how to structure this page.