RobustNeuralNetworks.AbstractREN

RobustNeuralNetworks.DiffREN

RobustNeuralNetworks.REN

RobustNeuralNetworks.WrapREN

Recurrent Equilibrium Networks (REN)

RobustNeuralNetworks.AbstractREN — Type

abstract type AbstractREN end

Explicit parameterisation for recurrent equilibrium networks.

(m::AbstractREN)(xt::VecOrMat, ut::VecOrMat)

Call an AbstractREN model given internal states xt and inputs ut.

If arguments are matrices, each column must be a vector of states or inputs (allows batch simulations).

Examples

This example creates a contracting REN using ContractingRENParams and calls the model with some randomly generated inputs.

using Random
using RobustNeuralNetworks

# Setup
rng = MersenneTwister(42)
batches = 10
nu, nx, nv, ny = 4, 2, 20, 1

# Construct a REN
contracting_ren_ps = ContractingRENParams{Float64}(nu, nx, nv, ny; rng=rng)
ren = REN(contracting_ren_ps)

# Some random inputs
x0 = init_states(ren, batches; rng=rng)
u0 = randn(rng, ren.nu, batches)

# Evaluate the REN over one timestep
x1, y1 = ren(x0, u0)

println(round.(y1;digits=2))

# output

[-31.41 0.57 -0.55 -3.56 -35.0 -18.28 -25.48 -7.49 -4.14 15.31]

See also REN, WrapREN, and DiffREN.

RobustNeuralNetworks.DiffREN — Type

DiffREN(ps::AbstractRENParams{T}) where T

Construct a differentiable REN from its direct parameterisation.

DiffREN is an alternative to REN and WrapREN that computes the explicit parameterisation every time the model is called. This is slow and computationally inefficient. However, it can be trained with Flux.jl (unlike WrapREN) and does not need to re-created if the parameters are updated (unlike REN).

The key feature is that the ExplicitParams struct is never stored, so an instance of DiffREN never has to be mutated or re-defined after it is created, even when learnable parameters are updated.

See also AbstractREN, REN, and WrapREN.

RobustNeuralNetworks.REN — Type

REN(ps::AbstractRENParams{T}) where T

Construct a REN from its direct parameterisation.

This constructor takes a direct parameterisation of REN (eg: a GeneralRENParams instance) and converts it to a callable explicit parameterisation of the REN.

See also AbstractREN, WrapREN, and DiffREN.

RobustNeuralNetworks.WrapREN — Type

WrapREN(ps::AbstractRENParams{T}) where T

Construct REN wrapper from its direct parameterisation.

WrapREN is an alternative to REN that stores the AbstractRENParams and ExplicitParams within the same object. This means that a new REN object does not have to be created each time the parameters are updated. Explicit REN parameters must be updated by the user if the direct parameters have changed.

Note that WrapREN cannot be used with Flux.jl, since it relies on mutating the WrapREN instance.

Examples

In this example, we create a REN satisfying some generic behavioural constraints and demonstrate how to update the REN wrapper if model parameters are changed.

using LinearAlgebra
using Random
using RobustNeuralNetworks

# Setup
rng = MersenneTwister(42)
batches = 10
nu, nx, nv, ny = 4, 10, 20, 2

Q = Matrix{Float64}(-I(ny))
R = 0.1^2 * Matrix{Float64}(I(nu))
S = zeros(Float64, nu, ny)

# Construct a REN
ren_ps = GeneralRENParams{Float64}(nu, nx, nv, ny, Q, S, R; rng=rng)
ren = WrapREN(ren_ps)

# Some dummy inputs
x0 = init_states(ren, batches; rng=rng)
u0 = randn(rng, ren.nu, batches)

# Evaluate the REN over one timestep
x1, y1 = ren(x0, u0) 

# Update the model after changing a parameter
ren.params.direct.B2 .*= rand(rng, size(ren.params.direct.B2)...)
update_explicit!(ren)

println(round(ren.explicit.B2[1];digits=4))

See also AbstractREN, REN, and DiffREN.