RobustNeuralNetworks.jl Documentation

Bringing robust machine learning to Julia.

RobustNeuralNetworks.jl Documentation

Examples

Index

Docstrings

TODO:

Examples

Here's an example. You should be able to see the import statement of Random.

using Random
a = 1
b = 2*rand()
2a + b

2.532605288485002

Can we continue using variables from this example?

println(a+b)

1.5326052884850019

We can even make things look like the REPL.

julia> a = 11
julia> b = 22
julia> a + b3

We can even delay execution of an example over a few different example blocks. Start a for loop here...

for i in 1:3
    j = i^2

Then write something insightful and finish it below...

    println(j)
end

1
4
9

It's worth having a look at the @setup macro as well when you can. It will make it much easier to write examples that include a number of lines of setup which should be hidden. Having said that, it might be useful to show the reader how you set up the example!

Most of your examples should be written with the @jldoctest macro. I'll give it a go below, but have a look at how ControlSystems.jl does things too.

Example:

using Random
using RobustNeuralNetworks

batches = 50
nu, nx, nv, ny = 4, 10, 20, 2

contracting_ren_ps = ContractingRENParams{Float64}(nu, nx, nv, ny)
contracting_ren = REN(contracting_ren_ps)

x0 = init_states(contracting_ren, batches)
u0 = randn(contracting_ren.nu, batches)

x1, y1 = contracting_ren(x0, u0)  # Evaluates the REN over one timestep

println(size(y1))

# output

(2, 50)

Index

RobustNeuralNetworks.AbstractLBDN

RobustNeuralNetworks.AbstractREN

RobustNeuralNetworks.AbstractREN

RobustNeuralNetworks.AbstractRENParams

RobustNeuralNetworks.ContractingRENParams

RobustNeuralNetworks.ContractingRENParams

RobustNeuralNetworks.DiffREN

RobustNeuralNetworks.DiffREN

RobustNeuralNetworks.DirectParams

RobustNeuralNetworks.ExplicitParams

RobustNeuralNetworks.GeneralRENParams

RobustNeuralNetworks.LBFN

RobustNeuralNetworks.LipschitzRENParams

RobustNeuralNetworks.PassiveRENParams

RobustNeuralNetworks.REN

RobustNeuralNetworks.WrapREN

RobustNeuralNetworks.direct_to_explicit

RobustNeuralNetworks.init_states

RobustNeuralNetworks.set_output_zero!

RobustNeuralNetworks.update_explicit!

Docstrings

Work on the presentation of this a bit....

RobustNeuralNetworks.AbstractLBDN — Type

abstract type AbstractLBDN end

Parameterisation for Lipschitz-bounded deep networks.

RobustNeuralNetworks.AbstractREN — Type

abstract type AbstractREN end

Explicit parameterisation for recurrent equilibrium networks.

RobustNeuralNetworks.AbstractREN — Method

(m::AbstractREN)(xt::VecOrMat, ut::VecOrMat)

Call a REN model given internal states xt and inputs ut.

If arguments are matrices, each column must be a vector of states or inputs (allows batch simulations).

Examples

This example creates a contracting REN using ContractingRENParams and calls the model with some randomly generated inputs.

using Random
using RobustNeuralNetworks

# Setup
rng = MersenneTwister(42)
batches = 10
nu, nx, nv, ny = 4, 2, 20, 1

# Construct a REN
contracting_ren_ps = ContractingRENParams{Float64}(nu, nx, nv, ny; rng=rng)
ren = REN(contracting_ren_ps)

# Some dummy inputs
x0 = init_states(ren, batches; rng=rng)
u0 = randn(rng, ren.nu, batches)

# Evaluate the REN over one timestep
x1, y1 = ren(x0, u0)

println(round.(y1;digits=2))

See also REN, WrapREN, and DiffREN.

RobustNeuralNetworks.AbstractRENParams — Type

abstract type AbstractRENParams{T} end

Direct parameterisation for recurrent equilibrium networks.

RobustNeuralNetworks.ContractingRENParams — Method

ContractingRENParams{T}(nu, nx, nv, ny; <keyword arguments>) where T

Construct direct parameterisation of a contracting REN.

The parameters can be used to construct an explicit REN model that has guaranteed, built-in contraction properties.

Arguments

nu::Int: Number of inputs.

nx::Int: Number of states.

nv::Int: Number of neurons.

ny::Int: Number of outputs.

Keyword arguments

nl=Flux.relu: Static nonlinearity (eg: Flux.relu or Flux.tanh).

αbar::T=1: Upper bound on the contraction rate with ᾱ ∈ (0,1].

See DirectParams documentation for arguments init, ϵ, bx_scale, bv_scale, polar_param, D22_zero, rng.

See also GeneralRENParams, LipschitzRENParams, PassiveRENParams.

RobustNeuralNetworks.ContractingRENParams — Method

ContractingRENParams(nv, A, B, C, D; ...)

Alternative constructor for ContractingRENParams that initialises the REN from a stable discrete-time linear system with state-space model

\[\begin{align*} x_{t+1} &= Ax_t + Bu_t \\ y_t &= Cx_t + Du_t. \end{align*}\]

TODO: This method may be removed in a later edition of the package. TODO: Make compatible with αbar ≠ 1.0.

RobustNeuralNetworks.DiffREN — Type

mutable struct DiffREN <: AbstractREN

Wrapper for REN type which automatically re-computes explicit parameters every time the model is called.

This is slow, but is compatible with Flux.jl

RobustNeuralNetworks.DiffREN — Method

DiffREN(ps::AbstractRENParams{T}) where T

Construct a differentiable REN from its direct parameterisation.

DiffREN is an alternative to REN that computes the explicit parameterisation every time the model is called. This is slow and computationally inefficient. However, it can be used with Flux.jl just like any other Flux model to do machine learning.

The difference to REN and WrapREN is that the ExplicitParams struct is never stored, so an instance of DiffREN never has to be mutated or re-defined after it is created, even when learnable parameters are updated.

See also REN and WrapREN.

RobustNeuralNetworks.DirectParams — Method

DirectParams{T}(nu, nx, nv; <keyword arguments>) where T

Construct direct parameterisation for an (acyclic) recurrent equilibrium network.

This is typically used by higher-level constructors when defining a REN, which take the direct parameterisation and define rules for converting it to an explicit parameterisation. See for example GeneralRENParams.

Arguments

nu::Int: Number of inputs.

nx::Int: Number of states.

nv::Int: Number of neurons.

Keyword arguments

init=:random: Initialisation method. Options are:

:random: Random sampling for all parameters.

:cholesky: Compute X with cholesky factorisation of H, sets E,F,P = I.

polar_param::Bool=true: Use polar parameterisation to construct H matrix from X in REN parameterisation (recommended).

D22_free::Bool=false: Specify whether to train D22 as a free parameter (true), or construct it separately from X3, Y3, Z3 (false). Typically use D22_free = true only for a contracting REN.

D22_zero::Bool=false: Fix D22 = 0 to remove any feedthrough.

bx_scale::T=0: Set scale of initial state bias vector bx.

bv_scale::T=1: Set scalse of initial neuron input bias vector bv.

ϵ::T=1e-12: Regularising parameter for positive-definite matrices.

rng::AbstractRNG=Random.GLOBAL_RNG: rng for model initialisation.

See Revay et al. (2021) for parameterisation details.

See also GeneralRENParams, ContractingRENParams, LipschitzRENParams, PassiveRENParams.

RobustNeuralNetworks.ExplicitParams — Type

mutable struct ExplicitParams{T}

Explicit REN parameter struct.

These parameters define a recurrent equilibrium network with model inputs and outputs $u_t, y_t$, neuron inputs and outputs $v_t,w_t$, and states x_t.

\[\begin{equation*} \begin{bmatrix} x_{t+1} \\ v_t \\ y_t \end{bmatrix} = \begin{bmatrix} A & B_1 & B_2 \\ C_1 & D_{11} & D_{12} \\ C_2 & D_{21} & D_{22} \\ \end{bmatrix} \begin{bmatrix} x_t \\ w_t \\ u_t \end{bmatrix} + \begin{bmatrix} b_x \\ b_v \\ b_y \end{bmatrix} \end{equation*}\]

See Revay et al. (2021) for more details on explicit parameterisations of REN.

RobustNeuralNetworks.GeneralRENParams — Method

GeneralRENParams{T}(nu, nx, nv, ny, Q, S, R; <keyword arguments>) where T

Construct direct parameterisation of a REN satisfying general behavioural constraints.

Behavioural constraints are encoded by the matrices Q,S,R in an incremental Integral Quadratic Constraint (IQC). See Equation 4 of Revay et al. (2021).

Arguments

nu::Int: Number of inputs.

nx::Int: Number of states.

nv::Int: Number of neurons.

ny::Int: Number of outputs.

Q::Matrix{T}: IQC weight matrix on model outputs

S::Matrix{T}: IQC coupling matrix on model outputs/inputs

R::Matrix{T}: IQC weight matrix on model outputs

Keyword arguments

nl=Flux.relu: Static nonlinearity (eg: Flux.relu or Flux.tanh).

αbar::T=1: Upper bound on the contraction rate with ᾱ ∈ (0,1].

See DirectParams documentation for arguments init, ϵ, bx_scale, bv_scale, polar_param, rng.

See also ContractingRENParams, LipschitzRENParams, PassiveRENParams.

RobustNeuralNetworks.LBFN — Method

LBFN{T}(nu, nh, ny, γ; ...)

Constructor for an LBFN with nu inputs, nv outputs, and nh = [nh1, nh2,...] specifying the size of hidden layers. User-imposed Lipschitz bound γ has a default of 1.

RobustNeuralNetworks.LipschitzRENParams — Method

LipschitzRENParams(nu, nx, nv, ny, γ; <keyword arguments>) where T

Construct direct parameterisation of a REN with a Lipschitz bound of γ.

Arguments

nu::Int: Number of inputs.

nx::Int: Number of states.

nv::Int: Number of neurons.

ny::Int: Number of outputs.

γ::Number: Lipschitz upper bound.

Keyword arguments

nl=Flux.relu: Static nonlinearity (eg: Flux.relu or Flux.tanh).

αbar::T=1: Upper bound on the contraction rate with ᾱ ∈ (0,1].

See DirectParams documentation for arguments init, ϵ, bx_scale, bv_scale, polar_param, D22_zero, rng.

See also GeneralRENParams, ContractingRENParams, PassiveRENParams.

RobustNeuralNetworks.PassiveRENParams — Method

PassiveRENParams{T}(nu, nx, nv, ny; <keyword arguments>) where T

Construct direct parameterisation of a passive REN.

Arguments

nu::Int: Number of inputs.

nx::Int: Number of states.

nv::Int: Number of neurons.

ny::Int: Number of outputs.

Keyword arguments

ν::T=0: Passivity parameter. Use ν>0 for incrementally strictly input passive model, and ν == 0 for incrementally passive model.

nl=Flux.relu: Static nonlinearity (eg: Flux.relu or Flux.tanh).

αbar::T=1: Upper bound on the contraction rate with ᾱ ∈ (0,1].

See DirectParams documentation for arguments init, ϵ, bx_scale, bv_scale, polar_param, rng.

See also GeneralRENParams, ContractingRENParams, LipschitzRENParams.

RobustNeuralNetworks.REN — Method

REN(ps::AbstractRENParams{T}) where T

Construct a REN from its direct parameterisation.

This constructor takes a direct parameterisation of REN (eg: a GeneralRENParams instance) and converts it to a callable explicit parameterisation of the REN.

RobustNeuralNetworks.WrapREN — Method

WrapREN(ps::AbstractRENParams{T}) where T

Construct REN wrapper from its direct parameterisation.

WrapREN is an alternative to REN that stores the AbstractRENParams and ExplicitParams within the same object. This means that a new REN object does not have to be created each time the parameters are updated. Explicit REN parameters must be updated by the user if the direct parameters have changed.

Note that WrapREN cannot be used with Flux.jl, since it relies on mutating the WrapREN instance.

Examples

In this example, we create a REN satisfying some generic behavioural constraints and demonstrate how to update the REN wrapper if model parameters are changed.

using LinearAlgebra
using Random
using RobustNeuralNetworks

# Setup
rng = MersenneTwister(42)
batches = 10
nu, nx, nv, ny = 4, 10, 20, 2

Q = Matrix{Float64}(-I(ny))
R = 0.1^2 * Matrix{Float64}(I(nu))
S = zeros(Float64, nu, ny)

# Construct a REN
ren_ps = GeneralRENParams{Float64}(nu, nx, nv, ny, Q, S, R; rng=rng)
ren = WrapREN(ren_ps)

# Some dummy inputs
x0 = init_states(ren, batches; rng=rng)
u0 = randn(rng, ren.nu, batches)

# Evaluate the REN over one timestep
x1, y1 = ren(x0, u0) 

# Update the model after changing a parameter
ren.params.direct.B2 .*= rand(rng, size(ren.params.direct.B2)...)
update_explicit!(ren)

println(round(ren.explicit.B2[1];digits=4))

See also REN and DiffREN.

RobustNeuralNetworks.direct_to_explicit — Function

direct_to_explicit(ps::AbstractRENParams, return_h=false) where T

Convert direct parameterisation of RENs to explicit parameterisation.

Uses the parameterisation encoded in ps to construct an ExplicitParams object that naturally satisfies a set of user-defined behavioural constraints.

Arguments

ps::AbstractRENParams: Direct parameterisation with behavioural constraints to convert to an explicit parameterisation of REN (eg: GeneralRENParams).

return_h::Bool=false: Whether to return the H-matrix directly (see Revay et al. (2021)). Useful for debugging or model analysis. If false, function returns an object of type ExplicitParams{T}.

See also GeneralRENParams, ContractingRENParams, LipschitzRENParams, PassiveRENParams.

RobustNeuralNetworks.init_states — Method

init_states(m::AbstractREN, nbatches; rng=nothing)

Return matrix of (nbatches) state vectors of a REN initialised as zeros.

RobustNeuralNetworks.set_output_zero! — Method

set_output_zero!(m::AbstractREN)

Set output map of a REN to zero.

If the resulting model is called with x1,y = ren(x,u) then y = 0 for any x and u.

RobustNeuralNetworks.update_explicit! — Method

update_explicit!(m::WrapREN)

Update explicit model in WrapREN using the current direct parameters.

TODO:

Add a logo

Fill out this main documentation page

See ControlSystems.jl for a good example of how to structure this page.